Update to Files
Updated Colors for EGL.

The germany.txt is my random spacing fixes as I am anal as hell.

Events Germany is to fix the 35 event for EGL to work with the new commonwealth event.

the ideas.gfx is to fix the gfx for a national spirit being broken.

Added EGL names to countries and commonwealth event localization to events.