Updates bro
Updates to germany FT to fix a bunch of issues.

Fixes to german event to remove tooltip that doesn't exist.

Fixes to Norway to make it so they use the puppets right.

Removal of load_screen gfx and gui as they are the same as vanilla.

Added more localization text for German FT