Updates to some Vanilla Files
Removed a good amount of vanilla files or renamed them and remove vanilla things we didn't edit.

Mainly to make the mod less bloat as a good amount of files had a good amount of unneeded vanilla things we didn't edit as we only added things to them.

Also did some more DoD edits that I missed in the first pass over the files.