Good Updates
Removed Buildings.txt as you said we are gonna use vanilla till we figure it out.

Modified a lot of focuses for germany. Mainly to do with how they are available as many of them previously didn't work it both Germany and a puppet owned provinces that were required so it was just updating them so that works as well as fixing some so it doesn't lock you out in weird cases.

Updates to events to use subject_of instead of puppet_of to work better.

Added in a event so that if ireland joins the axis through the event from the focus it lets you know instead of just it happening.

A fixed typo in portugal events.

Changed norway cosmetic tag and added it to them from the focus.

Added Ireland event localization and localization for a vanilla event that we added a option to.