Updates
Added Idea to italy that makes it so they won't join Germany till they take the state Paris is in.

Edited German FT to make it so Pact of Steel can't be taken till they take the state Paris is in. Also edited a focus so it bypasses if France is at war with Germany, if france doesn't exist, or if france is a subject of germany.

Added localization for Italies new events.

Added in the gui for the political view that is setup to show all 10 Ideologies. Note: I did not edit a texture for it. The texture that needs to be edited is pol_view_bg.dds in gfx/interface.

remove supply_area_names_l_english.yml as it was the same as it was auto generated by Zolt and the new supply areas are already in DH_supply_area_names_l_english.yml.

The pol_piechart_bg.dds and countrypoliticsview.gui.Slider is the files needed to have a slider instead of a bigger political party view. Adding them just incase.