Event Updates
Fixes to a good amount of the added events for the Asia block.

Updated documentation in most events to be consistant accross files.

Did a quick update to the German tree to use more of a vanilla approach to removing the Focus from Spain.

Added localizaiton for the Xinjiang(Sinkiang) event.

Renamed China event file to china instead of ROC.