Updated history files.
Either I standardized the parties popularity as some used a different look then others or added the parties popularity from HOI3 that were missed.

Only a few we missed as they had different tags in HOI3 then HOI4.

Also added history file and countries file for the new SPA and added them to the country tags file for whenever Zolt gets the spanish stuff up and running. This is only the begining as I assume he has more plans for this.