Surrender Event
Added the Bitter Peace Surrender Event for Soviets surrendering to the Germans.

Also added the localization for it and added some new flags to other events to account for the new events.

Removed the rank_defines.lua and instead of having two luas just added it to the o1_dh.lua as well as changed opinions from -100-100 to -200-200 max values.

Minor fixes to other files included.