Localization Change.
Came up with the idea to change localization files for the events to be organized like the event files actually are.

Basically event localization is separated by country now so you know exactly were to look to edit a specific countries event localization.

The rest of the files is rename/cleanup of older files spacing and other small things.