Changes
Minor spacing removal in a few files.

Added the opinion modifier for democratic ideologies to other democratic ideologies.

Edited the events to getting Vichy France in a attempt to 1)give them all French cores 2)actually make it so you can switch to VIC from FRA as our events didn't include how vanilla did that in the event so this is a attempt to get it looking right haven't tested it but should work like vanilla did.

Edited the puppets in Germany and Japan history files. Not sure if this was needed but included it anyway just in case.