German FT Bypass's First Pass.
Did a quick pass on the german focus tree and added bypass's to the focuses that didn't have one that were about taking countries/states.

Note: The spanish focuses need a rewrite to deal with the new way we do the civil war sides eg two tags instead of a civil war tag. Also This is just a first pass so things might have been missed or some that we want bypasses on might not have them yet.